+++ src/plugin/fieldViews/CheckboxFieldView.ts
