+++ src/plugin/fieldViews/RichTextFieldView.ts
