+++ src/renderers/react/createReactElementSpec.tsx
@@ -3,9 +3,9 @@ import React from "react";
 import { render } from "react-dom";
 import { createElementSpec } from "../../elementSpec";
 import type { Renderer, Validator } from "../../elementSpec";
-import type { FieldNameToValueMap } from "../../fieldViews/helpers";
-import type { Consumer } from "../../types/Consumer";
-import type { FieldSpec } from "../../types/Element";
+import type { FieldNameToValueMap } from "../../plugin/fieldViews/helpers";
+import type { Consumer } from "../../plugin/types/Consumer";
+import type { FieldSpec } from "../../plugin/types/Element";
 import { ElementProvider } from "./ElementProvider";
 
 export const createReactElementSpec = <
